Included CulturalHelper logic in /mosaic
