+++ package.json
@@ -32,6 +32,7 @@
     "compression": "^1.0.3",
     "cors": "^2.5.2",
     "helmet": "^1.3.0",
+    "lodash": "^4.17.4",
     "loopback": "^3.0.0",
     "loopback-boot": "^2.6.5",
     "loopback-component-explorer": "^4.0.0",
